requirements.txt
python-telegram-bot==21.*
httpx
beautifulsoup4
lxml
python-dotenv
exa-py
playwright==1.48.0
